Improve Maintainability
Combine similar code
Extract window util
